pyproject.toml
[project]
name = "shogun-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.2",
    "jinja2>=3.1.6",
    "libtmux>=0.53.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.40.0",
]
